requirements.txt
pandas>=2.2.3
openai>=1.78.1
python-dotenv>=1.1.0